Typescript Conversion
