+++ server/types/json.d.ts
@@ -1,0 +1,4 @@
+declare module "*.json" {
+    const value: any;
+    export default value;
+}
